+++ tests/integration/test_token_client.py
@@ -13,17 +13,11 @@ from ..utils import AIRDROP_AMOUNT, OPTS, assert_valid_response
 @pytest.fixture(scope="module")
 def test_token(stubbed_sender_for_token, freeze_authority, test_http_client) -> Token:
     """Test create mint."""
-    initial_balance = test_http_client.get_balance(
-        stubbed_sender_for_token.pubkey()
-    ).value
-    resp = test_http_client.request_airdrop(
-        stubbed_sender_for_token.pubkey(), AIRDROP_AMOUNT
-    )
+    initial_balance = test_http_client.get_balance(stubbed_sender_for_token.pubkey()).value
+    resp = test_http_client.request_airdrop(stubbed_sender_for_token.pubkey(), AIRDROP_AMOUNT)
     assert_valid_response(resp)
     test_http_client.confirm_transaction(resp.value)
-    final_balance = test_http_client.get_balance(
-        stubbed_sender_for_token.pubkey()
-    ).value
+    final_balance = test_http_client.get_balance(stubbed_sender_for_token.pubkey()).value
     assert final_balance - initial_balance >= AIRDROP_AMOUNT
     expected_decimals = 6
     token_client = Token.create_mint(
@@ -53,25 +47,19 @@ def test_token(stubbed_sender_for_token, freeze_authority, test_http_client) -> 
 
 
 @pytest.fixture(scope="module")
-def stubbed_sender_token_account_pk(
-    stubbed_sender_for_token, test_token
-) -> Pubkey:  # pylint: disable=redefined-outer-name
+def stubbed_sender_token_account_pk(stubbed_sender_for_token, test_token) -> Pubkey:  # pylint: disable=redefined-outer-name
     """Token account for stubbed sender."""
     return test_token.create_account(stubbed_sender_for_token.pubkey())
 
 
 @pytest.fixture(scope="module")
-def stubbed_receiver_token_account_pk(
-    stubbed_receiver, test_token
-) -> Pubkey:  # pylint: disable=redefined-outer-name
+def stubbed_receiver_token_account_pk(stubbed_receiver, test_token) -> Pubkey:  # pylint: disable=redefined-outer-name
     """Token account for stubbed receiver."""
     return test_token.create_account(stubbed_receiver)
 
 
 @pytest.mark.integration
-def test_new_account(
-    stubbed_sender_for_token, test_http_client, test_token
-):  # pylint: disable=redefined-outer-name
+def test_new_account(stubbed_sender_for_token, test_http_client, test_token):  # pylint: disable=redefined-outer-name
     """Test creating a new token account."""
     token_account_pk = test_token.create_account(stubbed_sender_for_token.pubkey())
     resp = test_http_client.get_account_info(token_account_pk)
@@ -86,9 +74,7 @@ def test_new_account(
         and not account_data.delegated_amount
         and Pubkey(account_data.delegate) == Pubkey([0] * 31 + [0])
     )
-    assert not account_data.close_authority_option and Pubkey(
-        account_data.close_authority
-    ) == Pubkey([0] * 31 + [0])
+    assert not account_data.close_authority_option and Pubkey(account_data.close_authority) == Pubkey([0] * 31 + [0])
     assert not account_data.is_native_option and not account_data.is_native
     assert Pubkey(account_data.mint) == test_token.pubkey
     assert Pubkey(account_data.owner) == stubbed_sender_for_token.pubkey()
@@ -107,9 +93,7 @@ def test_new_associated_account(test_token):  # pylint: disable=redefined-outer-
 
 
 @pytest.mark.integration
-def test_get_account_info(
-    stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token
-):  # pylint: disable=redefined-outer-name
+def test_get_account_info(stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token):  # pylint: disable=redefined-outer-name
     """Test get token account info."""
     account_info = test_token.get_account_info(stubbed_sender_token_account_pk)
     assert account_info.is_initialized is True
@@ -125,9 +109,7 @@ def test_get_account_info(
 
 
 @pytest.mark.integration
-def test_get_mint_info(
-    stubbed_sender_for_token, freeze_authority, test_token
-):  # pylint: disable=redefined-outer-name
+def test_get_mint_info(stubbed_sender_for_token, freeze_authority, test_token):  # pylint: disable=redefined-outer-name
     """Test get token mint info."""
     mint_info = test_token.get_mint_info()
     assert mint_info.mint_authority == stubbed_sender_for_token.pubkey()
@@ -138,9 +120,7 @@ def test_get_mint_info(
 
 
 @pytest.mark.integration
-def test_mint_to(
-    stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token
-):  # pylint: disable=redefined-outer-name
+def test_mint_to(stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token):  # pylint: disable=redefined-outer-name
     """Test mint token to account and get balance."""
     expected_amount = 1000
     assert_valid_response(
@@ -210,9 +190,7 @@ def test_burn(
 
 
 @pytest.mark.integration
-def test_mint_to_checked(
-    stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token
-):  # pylint: disable=redefined-outer-name
+def test_mint_to_checked(stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token):  # pylint: disable=redefined-outer-name
     """Test mint token checked and get balance."""
     expected_amount = 1000
     mint_amount = 700
@@ -266,9 +244,7 @@ def test_transfer_checked(
 
 
 @pytest.mark.integration
-def test_burn_checked(
-    stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token
-):  # pylint: disable=redefined-outer-name
+def test_burn_checked(stubbed_sender_for_token, stubbed_sender_token_account_pk, test_token):  # pylint: disable=redefined-outer-name
     """Test burning tokens checked."""
     burn_amount = 500
     expected_decimals = 6
@@ -291,13 +267,9 @@ def test_burn_checked(
 
 
 @pytest.mark.integration
-def test_get_accounts(
-    stubbed_sender_for_token, test_token
-):  # pylint: disable=redefined-outer-name
+def test_get_accounts(stubbed_sender_for_token, test_token):  # pylint: disable=redefined-outer-name
     """Test get token accounts."""
-    resp = test_token.get_accounts_by_owner_json_parsed(
-        stubbed_sender_for_token.pubkey()
-    )
+    resp = test_token.get_accounts_by_owner_json_parsed(stubbed_sender_for_token.pubkey())
     assert_valid_response(resp)
     assert len(resp.value) == 2
     for resp_data in resp.value:
@@ -386,9 +358,7 @@ def test_approve_checked(
 
 
 @pytest.mark.integration
-def test_freeze_account(
-    stubbed_sender_token_account_pk, freeze_authority, test_token, test_http_client
-):  # pylint: disable=redefined-outer-name
+def test_freeze_account(stubbed_sender_token_account_pk, freeze_authority, test_token, test_http_client):  # pylint: disable=redefined-outer-name
     """Test freezing an account."""
     resp = test_http_client.request_airdrop(freeze_authority.pubkey(), AIRDROP_AMOUNT)
     assert_valid_response(resp)
@@ -399,9 +369,7 @@ def test_freeze_account(
     account_info = test_token.get_account_info(stubbed_sender_token_account_pk)
     assert account_info.is_frozen is False
 
-    freeze_resp = test_token.freeze_account(
-        stubbed_sender_token_account_pk, freeze_authority, opts=OPTS
-    )
+    freeze_resp = test_token.freeze_account(stubbed_sender_token_account_pk, freeze_authority, opts=OPTS)
     assert_valid_response(freeze_resp)
     test_http_client.confirm_transaction(freeze_resp.value)
     account_info = test_token.get_account_info(stubbed_sender_token_account_pk)
@@ -409,16 +377,12 @@ def test_freeze_account(
 
 
 @pytest.mark.integration
-def test_thaw_account(
-    stubbed_sender_token_account_pk, freeze_authority, test_token, test_http_client
-):  # pylint: disable=redefined-outer-name
+def test_thaw_account(stubbed_sender_token_account_pk, freeze_authority, test_token, test_http_client):  # pylint: disable=redefined-outer-name
     """Test thawing an account."""
     account_info = test_token.get_account_info(stubbed_sender_token_account_pk)
     assert account_info.is_frozen is True
 
-    thaw_resp = test_token.thaw_account(
-        stubbed_sender_token_account_pk, freeze_authority, opts=OPTS
-    )
+    thaw_resp = test_token.thaw_account(stubbed_sender_token_account_pk, freeze_authority, opts=OPTS)
     assert_valid_response(thaw_resp)
     test_http_client.confirm_transaction(thaw_resp.value)
     account_info = test_token.get_account_info(stubbed_sender_token_account_pk)
@@ -452,9 +416,7 @@ def test_close_account(
 
 
 @pytest.mark.integration
-def test_create_multisig(
-    stubbed_sender_for_token, stubbed_receiver, test_token, test_http_client
-):  # pylint: disable=redefined-outer-name
+def test_create_multisig(stubbed_sender_for_token, stubbed_receiver, test_token, test_http_client):  # pylint: disable=redefined-outer-name
     """Test creating a multisig account."""
     min_signers = 2
     multisig_pubkey = test_token.create_multisig(